requirements.txt
nfl_data_py
pandas
google-cloud-storage
google-cloud-bigquery
pandas-gbq
tabulate
